+++ test-resources/namespaced_kw.clj
@@ -1,0 +1,7 @@
+(ns namespaced-kw)
+
+(defn bar
+  "prints baz"
+  [{:keys [baz]
+    :or   {baz ::my-kw}}]
+  (println baz))